requirements.txt
############################
# 必裝：目前程式已知會用到的套件
############################

Flask               # app.py: from flask import ...
PyMySQL             # app.py: import pymysql
requests            # app.py: import requests

python-dotenv       # 對應 .env.example，用來讀取環境變數（DB 密碼、API KEY）

openai              # langchain_agent.py 很可能會用到的 LLM SDK
langchain           # langchain_agent.OutfitAIAgent 所需
tiktoken            # LangChain / OpenAI 計算 token 時會用到

############################
# 資料處理 / 分析（對應 Numpy / Pandas 單元）
############################

numpy
pandas

############################
# 影像處理：使用者上傳衣服照片（對應 OpenCV 單元）
############################

opencv-python       # 分析衣服主色、邊緣等特徵
Pillow              # 圖片縮圖、格式轉換，比較輕量

############################
# 爬蟲 / 抓合作商家商品（對應爬蟲單元）
############################

beautifulsoup4      # 解析 HTML
lxml                # 加速 BeautifulSoup 解析
selenium            # 若要處理動態網頁（可選，看你有沒有實作）

############################
# Web 表單 / 前端體驗（可選，但對「登入/上傳/評分表單」很實用）
############################

Flask-WTF           # 穿搭評分表單、衣服上傳表單
Flask-Bootstrap     # 用 Bootstrap 美化畫面（說明你有運用前端框架）

############################
# 機器學習 / 推薦排序（對應「機器學習」單元）
############################

scikit-learn        # 例如根據評分歷史，調整推薦優先順序
matplotlib          # 畫圖展示推薦效果、評分分布（報告很好用）

############################
# 部署相關（可選，加分用）
############################

gunicorn            # 之後如果要用 Docker 部署 Flask，可用的 WSGI server
############################
# Web 後端框架（對應課綱 Flask 單元 + Jinja2 模板） 
############################
Flask
Flask-Bootstrap      # 簡單用 Bootstrap 美化頁面（加分：有用到前端 CSS/Bootstrap）
Flask-WTF            # 如果你有做表單（登入 / 上傳衣服資料等）
Jinja2               # 雖然 Flask 會自帶，但顯示出你有在用模板引擎
python-dotenv        # 管理 API key（例如 LLM、資料庫密碼）放在 .env 檔

############################
# 資料庫 / MySQL（對應課綱「關聯式資料庫 MySQL」）
############################
PyMySQL              # 你 app.py 裡的 import pymysql 對應的套件

############################
# 資料處理 / 機器學習（對應 Numpy / Pandas / Matplotlib / 機器學習 單元）
############################
numpy
pandas
scikit-learn         # 用來做簡單推薦排序、根據評分調整優先順序
matplotlib           # 做一些分析圖（顏色分布、評分分布），報告時很好展示

############################
# 影像處理（對應 OpenCV 影像辨識單元）
############################
opencv-python        # 抓衣服主色、簡單特徵用得到
Pillow               # 圖片縮圖、基本處理（有時比 OpenCV 好用）

############################
# 爬蟲 / 串接外部商品（對應網路爬蟲單元）
############################
requests             # 呼叫 LLM API、抓合作商家商品資料
beautifulsoup4       # 如果要從網頁 HTML 抓商品資訊
lxml                 # 搭配 bs4，速度比較快
selenium             # 若商家網站需要動態載入時可用（有學過就列進來）

############################
# LLM / AI Agent（你現在已經在用外部 LLM）
############################
openai               # 如果你是用 OpenAI / Azure OpenAI 之類的 LLM
langchain            # 你的 langchain_agent.py、OutfitAIAgent 用得到
tiktoken             # 給 langchain / openai 做 token 計算用

############################
# 部署 / 其他（可選，看你有沒有要用）
############################
gunicorn             # 之後若要部署（Heroku / Docker）可以用的 WSGI server
langchain==0.3.26
langchain-core
langchain-community
langchain-google-genai
chromadb
sentence-transformers
faiss-cpu
openai
tiktoken
python-dotenv
requests
transformers
numpy
scikit-learn
langgraph
langchain_chroma
